Re-organize the folder structure and merge the components and lib folders.
