+++ app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,3 @@
-import { handlers } from '@/app/lib/auth'; // Referring to the auth.ts we just created
+import { handlers } from '@/lib/auth';
+
 export const { GET, POST } = handlers;
