+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import localFont from 'next/font/local';
 
 import './globals.css';
-import { ThemeProvider } from './components/ThemeProvider';
+import { ThemeProvider } from '../components/custom/ThemeProvider';
 
 const geistSans = localFont({
   weight: '100 900',
