+++ components/custom/Navbar.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link';
 import Image from 'next/image';
 
-import Logo from '@/public/logo.png';
 import AuthModal from './AuthModal';
+import Logo from '@/public/logo.png';
 
 const Navbar = () => {
   return (
