+++ lib/auth.ts
@@ -5,6 +5,6 @@ import Google from 'next-auth/providers/google';
 import { PrismaAdapter } from '@auth/prisma-adapter';
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
-  adapter: PrismaAdapter(prisma),
   providers: [GitHub, Google],
+  adapter: PrismaAdapter(prisma),
 });
