+++ lib/hooks.ts
@@ -1,4 +1,5 @@
 import { redirect } from 'next/navigation';
+
 import { auth } from './auth';
 
 export async function requireUser() {
